+++ R/fw_model.R
@@ -8,31 +8,34 @@
 #' @param B biomass vector.
 #'
 #' @details
-#' So far there is only one model supported, a general Lotka-Volterra with 
-#' a linear response. 
-#' 
+#' So far there is only one model supported, a general Lotka-Volterra with
+#' a linear response.
+#'
 #' @return
 #' An object of class `fw_problem` which is a named list of 4 elements:
 #' * A: interaction matrix
 #' * B: biomass vector
 #' * R: mortality rate vector
 #' * model: the underlying model
-#' 
+#'
 #' @export
-#' 
+#'
 #' @examples
 #' A <- rbind(c(-1, -1), c(1, 0))
 #' R <- c(0.1, -0.05)
 #' B <- c(0.5, 0.25)
 #' fw_model(A, B, R)
-
 fw_model <- function(A, B, R) {
     x <- new_fw_model(A, B, R)
     validate_fw_model(x)
+    x$leading_ev <- get_leading_ev(
+        get_jacobian(x$A, x$B, x$R, mod = x$model)
+    )
+    x
 }
 
 #' Create models
-#' @describeIn fw_model get model 
+#' @describeIn fw_model get model
 #' @param model model to be used.
 #' @export
 
@@ -57,16 +60,17 @@ new_fw_model <- function(A, B, R) {
             A = A,
             B = B,
             R = R,
-            model = fw_get_model()
+            model = fw_get_model(),
+            leading_ev = NA
         ),
         class = "fw_model"
     )
 }
 
 validate_fw_model <- function(x) {
-    A <- x$A   
-    B <- x$B   
-    R <- x$R  
+    A <- x$A
+    B <- x$B
+    R <- x$R
     stopifnot(exprs = {
         inherits(A, "matrix")
         NROW(A) == NCOL(A)